Add files via upload
This dashboard helps us to analyze and track the open vulnerabilities in the environments. It compares the current month's vulnerability report with the previous month's report and also shows the KPI and provides reports for the last months for which data was uploaded. It also has a write-back feature added using Write! edit console where CSV files are used as storage to track the current status of the vulnerabilities and get the analysis out of it.