Add files via upload
This is a simple sales validation dashboard which helps team to check and verify how much sales has been done for every country. The source of this dashboard is ERP system.
This dashboard saves 90% of teams efforts of checking and validating data manually from the source. Also, helps team to take preventive actions in case of any issues and before the actual country sales dashboard chains gets completed.