Add files via upload
Maintaining the Qlik Sense is service cluster disk storage is essential as we don't want it to get filled with unused files and unused and not updated Qlik Sense QVF files are the one of the thing that we can consider for cleanup.
This is a customized Qlik Sense dashboard which is useful to track the number of unpublished applications under users workspace which are not reloading since some time.
It provides a detailed overview of how many such apps are there and under who's name. Also, the dashboard is integrated with mail and deploy for auto communication to those users with the details so that they can take of such apps.